Make the puppetmaster shell provisioner a little more idempotent.

+++ Vagrantfile
@@ -141,12 +141,17 @@ Vagrant::Config.run do |config|
 
     # set up all puppet nodes to be an apache web server
     vm_config.vm.provision :shell, :inline => <<-PREPARE_MASTER.gsub(/^ {6}/, '')
-      puppet module install puppetlabs-apache
-      cat <<SITE_PP > /etc/puppet/manifests/site.pp
+      if ! puppet module list | grep -q puppetlabs-apache >/dev/null ; then
+        puppet module install puppetlabs-apache
+      fi
+
+      if [ ! -f /etc/puppet/manifests/site.pp ] ; then
+        cat <<SITE_PP > /etc/puppet/manifests/site.pp
       node default {
         class { 'apache': }
       }
       SITE_PP
+      fi
     PREPARE_MASTER
   end
